requirements.txt
# ============================================================
#  Smart Agriculture Data Platform — Python Dependencies
# ============================================================

# ── Google Earth Engine ──────────────────────────────────────
earthengine-api>=0.1.390
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# ── Copernicus CDS (ERA5 / Climate) ─────────────────────────
cdsapi>=0.6.1

# ── NASA AppEEARS / HTTP APIs ────────────────────────────────
requests>=2.31.0
pydap>=3.4.0

# ── SMAP / HDF5 ─────────────────────────────────────────────
h5py>=3.10.0

# ── Geospatial ───────────────────────────────────────────────
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.2
pyproj>=3.6.1
fiona>=1.9.5

# ── NetCDF / xarray ─────────────────────────────────────────
netCDF4>=1.6.5
xarray>=2023.12.0
dask[complete]>=2023.12.0
zarr>=2.16.1

# ── Data Science ─────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
click>=8.1.7
schedule>=1.2.1
tqdm>=4.66.0
loguru>=0.7.2
tenacity>=8.2.3
pyyaml>=6.0.1
